+++ app/src/main/java/infinumacademy/showsapp/kristinakoneva/showdetailsscreen/ShowDetailsViewModel.kt
@@ -1,4 +1,4 @@
-package infinumacademy.showsapp.kristinakoneva.show_details_screen
+package infinumacademy.showsapp.kristinakoneva.showdetailsscreen
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
