+++ app/src/main/java/infinumacademy/showsapp/kristinakoneva/showsscreen/ShowsAdapter.kt
@@ -1,4 +1,4 @@
-package infinumacademy.showsapp.kristinakoneva.shows_screen
+package infinumacademy.showsapp.kristinakoneva.showsscreen
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
