+++ app/src/main/java/infinumacademy/showsapp/kristinakoneva/showsscreen/ShowsViewModelFactory.kt
@@ -1,4 +1,4 @@
-package infinumacademy.showsapp.kristinakoneva.shows_screen
+package infinumacademy.showsapp.kristinakoneva.showsscreen
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
