+++ app/src/main/res/navigation/main.xml
@@ -8,7 +8,7 @@
 
     <fragment
         android:id="@+id/loginFragment"
-        android:name="infinumacademy.showsapp.kristinakoneva.login_screen.LoginFragment"
+        android:name="infinumacademy.showsapp.kristinakoneva.loginscreen.LoginFragment"
         android:label="LoginFragment"
         tools:layout="@layout/fragment_login"
         >
@@ -39,7 +39,7 @@
     </fragment>
     <fragment
         android:id="@+id/registerFragment"
-        android:name="infinumacademy.showsapp.kristinakoneva.register_screen.RegisterFragment"
+        android:name="infinumacademy.showsapp.kristinakoneva.registerscreen.RegisterFragment"
         android:label="RegisterFragment"
         tools:layout="@layout/fragment_register">
         <action
